refactor(eval): address review — per-subcommand dirs, type aliases, drop --yes
- Split llm-as-a-judge and code-based create/update into their own
  directories per the documented handler convention; shared flags/helpers
  live in a sibling utils.tsx.
- LlmAsAJudgeUpdate / CodeBasedUpdate are now type aliases (data-carrying),
  leaving CoreEvalClient as the implemented interface.
- Drop --yes from evaluator delete to stay consistent with the other CRUDL
  commands.
