+++ src/handlers/eval/evaluator/llm-as-a-judge/update/index.tsx
@@ -1,0 +1,41 @@
+import z from "zod";
+import { createHandler, flag } from "../../../../../router";
+import { JsonRendererKey } from "../../../../../tui";
+import type { AppIO, Core } from "../../../../types";
+import { coreOptsFromCtx } from "../../../../utils";
+import { SourceResolver } from "../../../source";
+import { instructionsFlag, ratingScaleFlag, resolveRatingScale } from "../utils";
+
+export const createLlmAsAJudgeUpdateHandler = (core: Core, io: AppIO) =>
+  createHandler({
+    name: "update",
+    description: "update an LLM-as-a-Judge evaluator",
+    flags: [
+      flag("id", "the ID of the evaluator to update", z.string().optional()),
+      instructionsFlag,
+      flag("model", "the Bedrock model id used to judge", z.string().optional()),
+      ratingScaleFlag,
+      flag("kms-key-arn", "customer managed KMS key ARN for evaluator data", z.string().optional()),
+      flag("client-token", "idempotency token", z.string().optional()),
+    ],
+    handle: async (ctx, flags) => {
+      if (!flags["id"]) throw new TypeError("required option '--id <id>' not specified");
+
+      const source = new SourceResolver(io);
+      const instructions = await source.resolve("instructions", flags["instructions"]);
+      const ratingScale = await resolveRatingScale(flags["rating-scale"], source);
+
+      const response = await core.eval.updateLlmAsAJudgeEvaluator(
+        flags["id"],
+        {
+          instructions,
+          model: flags["model"],
+          ratingScale,
+          kmsKeyArn: flags["kms-key-arn"],
+          clientToken: flags["client-token"],
+        },
+        coreOptsFromCtx(ctx),
+      );
+      ctx.require(JsonRendererKey).renderJson(response);
+    },
+  });
