+++ src/handlers/eval/types.tsx
@@ -14,23 +14,23 @@ import type { CoreOptions } from "../../core/types";
 // the AgentCore UpdateEvaluator API replaces the whole evaluatorConfig union, and
 // the llmAsAJudge arm requires instructions + ratingScale + modelConfig together,
 // so a partial update is only possible by merging over the current definition.
-export interface LlmAsAJudgeUpdate {
+export type LlmAsAJudgeUpdate = {
   instructions?: string;
   model?: string;
   ratingScale?: RatingScale;
   kmsKeyArn?: string;
   clientToken?: string;
-}
+};
 
 // CodeBasedUpdate carries the fields a caller may change on a code-based
 // evaluator. Undefined fields are preserved from the existing evaluator, for the
 // same union-replacement reason described on LlmAsAJudgeUpdate.
-export interface CodeBasedUpdate {
+export type CodeBasedUpdate = {
   lambdaArn?: string;
   timeout?: number;
   kmsKeyArn?: string;
   clientToken?: string;
-}
+};
 
 // CoreEvalClient is the evaluator surface the eval handlers depend on. It is
 // declared here, next to the handlers that consume it, and implemented by
